Netflix Movie Recommendation Program
This program acts in a similar way a subscription movie service does, as it takes the movies.txt file and splits the titles and descriptions into a list. Then the number of movies is taken and another list is created to store the similarity values. The movies are then iterated through and the similarity between the movie descriptions is checked with the most recently watched movie. The max similarity value is then taken, the most similar movie is returned and the user is given their recommendation.

semantic.py is simply a program that demonstrates the difference of using a multi data model (en_core_web_md) compared to en_core_web_sm, which is far more limited in its results.